feat: Refactoring for common queue objects
