+++ programs/match_template/run_distributed_match_template.py
@@ -71,7 +71,7 @@ def initialize_distributed() -> tuple[int, int, int]:
     default=ORIENTATION_BATCH_SIZE,
     help="Number of orientations to process in a single batch.",
 )
-def main() -> None:
+def main(config: str, output: str, batch_size: int) -> None:
     """Main function for the distributed match_template program.
 
     Each process is associated with a single GPU, and we front-load the distributed
@@ -79,19 +79,33 @@ def main() -> None:
     object and the backend match_template code to remain relatively simple.
     """
     world_size, rank, local_rank = initialize_distributed()
-    print(f"RANK={rank}: Initialized {world_size} processes (local_rank={local_rank}).")
+    time_str = time.strftime("%Y%m%d_%H%M%S", time.localtime())
+    print(
+        f"{time_str} RANK={rank}: Initialized {world_size} processes "
+        f"(local_rank={local_rank})."
+    )
 
     # Do not pre-load mrc files, unless zeroth rank. Data will be broadcast later.
     mt_manager = MatchTemplateManager.from_yaml(
-        YAML_CONFIG_PATH, preload_mrc_files=bool(rank == 0)
+        config, preload_mrc_files=bool(rank == 0)
     )
     mt_manager.run_match_template_distributed(
         world_size=world_size,
         rank=rank,
         local_rank=local_rank,
-        orientation_batch_size=ORIENTATION_BATCH_SIZE,
+        orientation_batch_size=batch_size,
         do_result_export=(rank == 0),  # Only save results from rank 0
     )
+
+    # Only do the df export on rank 0
+    if rank == 0:
+        df = mt_manager.results_to_dataframe()
+        df.to_csv(output, index=True)
+
+    # Close the distributed process group
+    dist.destroy_process_group()
+
+    print("Done!")
 
 
 if __name__ == "__main__":
